Teste Fullstack/apitestefs/requirements.txt
django
djangorestframework
environs
psycopg2
django-cors-headers==4.3.1